Migrate all headers to the contest template dialect
- rep/sz/pii/vpi/trav and Benq macros (each, ckmin/ckmax, V<>, AR<>, str,
  rsz, eb, bk/ft, re/ps) replaced with FOR/F0R/ROF, size(), pi, vt<pi>,
  range-for, std names; vector<int>/vector<ll>/vector<T> -> vi/vl/vt<T>,
  pair<int,int>/pair<ll,ll> -> pi/pl, long long -> ll, double -> db,
  push_back -> pb; duplicate local typedefs removed
- remaining aggregate Point ctor calls brace-initialised
  (linearTransformation, FastDelaunay)
- 161/191 headers now compile against template.cpp (was 139); the rest
  are intentional snippets/scaffolds, extc++-ordering files, or the
  standalone Benq programs
- stress tests: 95/119 pass; repaired test-side .size()/sz() uses
  (AhoCorasick, SuffixArray, CircleTangents, LinkCutTree); the 24 that
  still fail are pre-existing stale harnesses (old API names / Point
  ctor syntax), unrelated to the migration

+++ content/various/Timer.h
@@ -17,9 +17,9 @@ struct Timer {
   Timer() { reset(); }
   void reset() { st = chrono::high_resolution_clock::now(); }
 
-  long long elapsed() {
+  ll elapsed() {
     auto ed = chrono::high_resolution_clock::now();
     return chrono::duration_cast<chrono::milliseconds>(ed - st).count();
   }
-  long long operator()() { return elapsed(); }
+  ll operator()() { return elapsed(); }
 };
